feat:增加根据模板创建调查问卷功能

+++ src/main/java/com/gyb/questionnaire/entity/Template.java
@@ -22,6 +22,6 @@ public class Template {
     private java.sql.Timestamp publishDate;
     private long authorId;
     private String authorName;
-    private long typeId;
+    private int typeId;
     private int questionCount;
 }
